requirements.txt
Appium-Python-Client==1.2.0
appium-selector==1.3.2
Appium-UIAutomation==0.1.0
AppiumAndroid==0.1.2
AppiumRunner==0.0.1
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.10.0
by==0.0.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
cryptography==35.0.0
h11==0.12.0
http-api==0.2.2
http-proxy==1.0.16
idna==3.3
iniconfig==1.1.1
outcome==1.1.0
packaging==21.3
pluggy==1.0.0
proxy==0.0.1
py==1.10.0
pycparser==2.21
PyHamcrest==2.0.2
pyOpenSSL==21.0.0
pyparsing==3.0.6
pytest==6.2.5
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-runner==5.3.1
pytest-variables==1.9.0
python-http-client==3.3.3
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
selenium==3.141.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
suds-sw==0.4.3
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
wsproto==1.0.0
xlrd==2.0.1
